riceprediction/requirement.txt
WTForms==2.3.3
flask
Flask-WTF
flask-migrate
flask-script
gunicorn
pymysql
email-validator
cryptography
bootstrap-flask
Flask-SQLAlchemy
flask-mail
biopython
matplotlib
pandas
cymysql
plotly
optuna
lightgbm
pyecharts